Separating RPC
